adding test for caching ctx
and making the ctx test driver generic
